Menu Raw Example
Given as an example of a menu returned from scraping. Note there are plenty of extra fields to pull that would also need sanitized by Sanitizemecaptain and the likes to bring in their data. THC & CBD info to be included as it's own table referenced via productID as it's the only way to key it correctly and accurately over time.